Add endpoints to change route status and vertex status.

+++ backend/DTO/RoadmapProgress.cs
@@ -3,7 +3,7 @@
     public class RoadmapProgress
     {
         public bool IsStarted {  get; set; }
-        public float Percent { get; set; }
+        public int Percent { get; set; }
         public int Finished { get; set; }
         public int InProgress { get; set; }
         public int Skipped { get; set; }
